Add files via upload
1051. Height Checker
Sorting the heights list, then you can get a new list sorted. By comparing the value in the  same index, you can get the result if it is different.
Drawback: it needs a few time when sorting

1122. Relative Sort Array
comparing the number of arr2 with arr1, if equals, you can put the number in arr1 to a new list. After that,  sorting the rest of the number in arr1,merge it with the new list.  it is the result

883. Projection Area of 3D Shapes
From the top, the shadow made by the shape will be 1 square for each non-zero value.
From the side, the shadow made by the shape will be the largest value for each row in the grid.
From the front, the shadow made by the shape will be the largest value for each column in the grid.  From the solution of leetcode